factorize code in skeletinization and segmentation
